Add dynamic text resizing which renders maxLength obsolete (for now only supports setting cells once = doesn't support setting optimal font size when setting shorter text to cell which previously had longest text)
# Thoughts about adjustFontSize (maxLength vs maximum set dynamically)

So setting the fontSize per cell means that different cells have different font sizes.
(for example cell with content 100 and 4096)

Plus setting the fontSize per individual cell has bad perfomance even if the inner while loop runs 0 ranTimes

Also, our optimal font size computation relies on the actual elements.

So, if have maxLength == 4, and we want to keep the same font size **from the beginning**, when inserted values are small, like 1, 5,
**to the end**, when the values are 1000, we **need to have a reference element which has 4 digits from the beginning**
=> if we use 4 nbsp spaces, that needs a monospace font to work, else the width of space and digit is different

So, if we don't have a monospace font and want to have the same font size from the beginning, we can't really do it

(aside from hacks like setting the reference element with 0000 in the .green wall or black background (not ∞ cells), with the digits of transparent color)
=> this would probably work, but require at least one wall or at least one black background (There are some edge cases + this is also quite application specific, and I want my grid to be quite general)

## => So, maybe it is easier to have DYNAMIC but same font size
- so, in the begining when only entering small values, the size would be bigger
- then, as a larger element is added, it will become the reference element and the font size of all elements will be altered to its font size

- or alternatively make it go out of sync with adjustFontSize being adjusted per cell on setTextToCell and having horrendously slow resizes (with the for loop = without just setting the font size)

### TLDR: SO I WILL NOW IMPLEMENT THE APPROACH WITH THE DYNAMIC FONT SIZE (BUT SAME FOR ALL ELEMENTS), because it doesn't need any hacks => is cleaner RN

The implementation I'm going to do now is incomplete because it only supports adding longer and longer text and not shorter
(For arbitrary operations, it is needed to maintain a list of longest texts for tracking them => an ordered map = a tree)
(because it is possible, that shorter text will be set to the cell which has longest text right now => some other cell will have the longest text)